+++ src/fauxmoESP.h
@@ -1,8 +1,8 @@
 /*
 
-FAUXMO ESP 2.4.3
+FAUXMO ESP 3.0.0
 
-Copyright (C) 2016 by Xose Pérez <xose dot perez at gmail dot com>
+Copyright (C) 2016-2018 by Xose Pérez <xose dot perez at gmail dot com>
 
 The MIT License (MIT)
 
@@ -28,26 +28,19 @@ THE SOFTWARE.
 
 #pragma once
 
-#define DEFAULT_TCP_BASE_PORT   52000
 #define UDP_MULTICAST_IP        IPAddress(239,255,255,250)
 #define UDP_MULTICAST_PORT      1900
 #define TCP_MAX_CLIENTS         10
-
-#define UDP_SEARCH_PATTERN      "M-SEARCH"
-#define UDP_DEVICE_PATTERN_1    "urn:Belkin:device:**"
-#define UDP_DEVICE_PATTERN_2    "urn:Belkin:device:controllee:1"
-#define UDP_DEVICE_PATTERN_3    "urn:Belkin:service:basicevent:1"
-#define UDP_DEVICE_PATTERN_4    "ssdp:all"
-#define UDP_DEVICE_PATTERN_5    "ssdpsearch:all"
-#define UDP_ROOT_DEVICE         "upnp:rootdevice"
-
-#define UDP_RESPONSES_INTERVAL  250
-#define UDP_RESPONSES_TRIES     5
+#define TCP_PORT                1901
 
 #ifdef DEBUG_FAUXMO
     #define DEBUG_MSG_FAUXMO(...) DEBUG_FAUXMO.printf( __VA_ARGS__ )
 #else
     #define DEBUG_MSG_FAUXMO(...)
+#endif
+
+#ifndef DEBUG_FAUXMO_VERBOSE
+#define DEBUG_FAUXMO_VERBOSE    false
 #endif
 
 #include <Arduino.h>
@@ -65,42 +58,31 @@ THE SOFTWARE.
 #include <WiFiUdp.h>
 #include <functional>
 #include <vector>
-#include <WeMo.h>
+#include <templates.h>
 
-typedef std::function<void(unsigned char, const char *, bool)> TSetStateCallback;
-typedef std::function<bool(unsigned char, const char *)> TGetStateCallback;
+typedef std::function<void(unsigned char, const char *, bool, unsigned char)> TSetStateCallback;
 
 typedef struct {
     char * name;
-    char * uuid;
-	char * serial;
-    bool hit;
     bool state;
-    AsyncServer * server;
+    unsigned char value;
 } fauxmoesp_device_t;
 
 class fauxmoESP {
 
     public:
 
-        fauxmoESP(unsigned int port = DEFAULT_TCP_BASE_PORT);
-
         unsigned char addDevice(const char * device_name);
         bool renameDevice(unsigned char id, const char * device_name);
         char * getDeviceName(unsigned char id, char * buffer, size_t len);
         void onSetState(TSetStateCallback fn) { _setCallback = fn; }
-        void onGetState(TGetStateCallback fn) { _getCallback = fn; }
         void enable(bool enable);
         void handle();
 
-        // backwards compatibility DEPRECATED
-        void onMessage(TSetStateCallback fn) { onSetState(fn); }
-        void setState(unsigned char id, bool state);
-
     private:
 
+        AsyncServer * _server;
         bool _enabled = true;
-        unsigned int _base_port = DEFAULT_TCP_BASE_PORT;
         std::vector<fauxmoesp_device_t> _devices;
 		#ifdef ESP8266
         WiFiEventHandler _handler;
@@ -108,26 +90,18 @@ class fauxmoESP {
         WiFiUDP _udp;
         AsyncClient * _tcpClients[TCP_MAX_CLIENTS];
         TSetStateCallback _setCallback = NULL;
-        TGetStateCallback _getCallback = NULL;
 
-        unsigned int _roundsLeft = 0;
-        unsigned int _current = 0;
-        unsigned long _lastTick;
-        IPAddress _remoteIP;
-        unsigned int _remotePort;
-        unsigned int _udpPattern;
-        bool _discovering = false;
+        String _deviceJson(unsigned char id);
 
-        void _sendUDPResponse(unsigned int device_id);
-        void _nextUDPResponse();
-
-        void _handleSetup(AsyncClient *client, unsigned int device_id, void *data, size_t len);
-        void _handleMetaInfoService(AsyncClient *client, unsigned int device_id, void *data, size_t len);
-        void _handleEventService(AsyncClient *client, unsigned int device_id, void *data, size_t len);
-        void _handleControl(AsyncClient *client, unsigned int device_id, void *data, size_t len);
-
+        void _handleUDP();
         void _onUDPData(const IPAddress remoteIP, unsigned int remotePort, void *data, size_t len);
-        void _onTCPData(AsyncClient *client, unsigned int device_id, void *data, size_t len);
-        void _onTCPClient(AsyncClient *client, unsigned int device_id);
+        void _sendUDPResponse();
+
+        void _onTCPClient(AsyncClient *client);
+        void _onTCPData(AsyncClient *client, void *data, size_t len);
+        void _onTCPDescription(AsyncClient *client, void *data, size_t len);
+        void _onTCPList(AsyncClient *client, void *data, size_t len);
+        void _onTCPControl(AsyncClient *client, void *data, size_t len);
+        void _sendTCPResponse(AsyncClient *client, const char * code, char * body, const char * mime);
 
 };