+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0] 2018-08-15
+### Changed
+- Completely new paradigm, emulating Philips Hue API instead of Belkin Wemo API
+- onSetState signature has changed to accommodate the brightness value
+- onGetState no longer exists
+
+### Added
+- DEBUG_FAUXMO_VERBOSE setting for verbose debug output (defaults to false)
+
 ## [2.4.3] 2018-06-23
 Dummy version to force PlatformIO parsing